% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accentless}
\alias{accentless}
\title{Internal Function to Remove Accents}
\usage{
accentless(x)
}
\arguments{
\item{x}{A \code{character} vector.}
}
\value{
A A \code{character} vector without accents.
}
\description{
Internal Function to Remove Accents
}
\keyword{internal}
