% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idex.R
\name{idex_ifra}
\alias{idex_ifra}
\title{Idex-Ifra Database}
\usage{
idex_ifra(index = c("geral", "core"))
}
\arguments{
\item{index}{The index to be called. One of: \code{geral}, \code{core}.}
}
\value{
A \code{tibble} with \code{12} columns.
}
\description{
JGP Corporate Debt Securities Infrastructure Index.
}
\examples{
if (FALSE) {
  idex_ifra("core")
}
}
