% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idex.R
\name{idex_cdi}
\alias{idex_cdi}
\title{Idex-CDI Database}
\usage{
idex_cdi(index = c("geral", "core", "low_rated"))
}
\arguments{
\item{index}{The index to be called. One of: \code{geral}, \code{core} or \code{low_rated}.}
}
\value{
A \code{tibble} with \code{12} columns.
}
\description{
JGP Corporate Debt Securities CDI Index.
}
\examples{
if (FALSE) {
  idex_cdi("geral")
}
}
